fix: add shared/ to lint, format, and lint-staged
