+++ shared/assets/build.ts
@@ -9,9 +9,9 @@
 import * as fs from 'fs';
 import * as path from 'path';
 
-import type { CatalogEntry } from './types.js';
 import type { FurnitureManifest, InheritedProps, ManifestGroup } from './manifestUtils.js';
 import { flattenManifest } from './manifestUtils.js';
+import type { CatalogEntry } from './types.js';
 
 // ── Furniture catalog ─────────────────────────────────────────────────────────
 
@@ -35,10 +35,14 @@ export function buildFurnitureCatalog(assetsDir: string): CatalogEntry[] {
       if (manifest.type === 'asset') {
         // Single-asset manifest — validate required fields
         if (
-          manifest.width == null ||
-          manifest.height == null ||
-          manifest.footprintW == null ||
-          manifest.footprintH == null
+          manifest.width === undefined ||
+          manifest.width === null ||
+          manifest.height === undefined ||
+          manifest.height === null ||
+          manifest.footprintW === undefined ||
+          manifest.footprintW === null ||
+          manifest.footprintH === undefined ||
+          manifest.footprintH === null
         ) {
           continue;
         }
