+++ shared/assets/colorUtils.ts
@@ -1,0 +1,13 @@
+/**
+ * Pure color conversion utilities — no external dependencies.
+ */
+
+import { PNG_ALPHA_THRESHOLD } from './constants.js';
+
+export function rgbaToHex(r: number, g: number, b: number, a: number): string {
+  if (a < PNG_ALPHA_THRESHOLD) return '';
+  const rgb =
+    `#${r.toString(16).padStart(2, '0')}${g.toString(16).padStart(2, '0')}${b.toString(16).padStart(2, '0')}`.toUpperCase();
+  if (a >= 255) return rgb;
+  return `${rgb}${a.toString(16).padStart(2, '0').toUpperCase()}`;
+}
