+++ shared/assets/pngDecoder.ts
@@ -7,13 +7,13 @@
 
 import { PNG } from 'pngjs';
 
+import { rgbaToHex } from './colorUtils.js';
 import {
   CHAR_FRAME_H,
   CHAR_FRAME_W,
   CHAR_FRAMES_PER_ROW,
   CHARACTER_DIRECTIONS,
   FLOOR_TILE_SIZE,
-  PNG_ALPHA_THRESHOLD,
   WALL_BITMASK_COUNT,
   WALL_GRID_COLS,
   WALL_PIECE_HEIGHT,
@@ -21,16 +21,6 @@ import {
 } from './constants.js';
 import type { CharacterDirectionSprites } from './types.js';
 export type { CharacterDirectionSprites } from './types.js';
-
-// ── Pixel conversion ─────────────────────────────────────────
-
-export function rgbaToHex(r: number, g: number, b: number, a: number): string {
-  if (a < PNG_ALPHA_THRESHOLD) return '';
-  const rgb =
-    `#${r.toString(16).padStart(2, '0')}${g.toString(16).padStart(2, '0')}${b.toString(16).padStart(2, '0')}`.toUpperCase();
-  if (a >= 255) return rgb;
-  return `${rgb}${a.toString(16).padStart(2, '0').toUpperCase()}`;
-}
 
 // ── Sprite decoding ──────────────────────────────────────────
